api: make limit,offset int64

+++ api/context.go
@@ -29,7 +29,7 @@ func IDFromContext(ctx context.Context) (token string, err error) {
 	return IDFromMetadata(md)
 }
 
-func LimitAndOffsetFromContext(ctx context.Context) (limit, offset int, err error) {
+func LimitAndOffsetFromContext(ctx context.Context) (limit, offset int64, err error) {
 	md, err := MetadataFromContext(ctx)
 	if err != nil {
 		return 0, 0, err
